Issue 145: Create "Persistence" port Reorganize query functions and repositories in "Doctrine" and "InMemory" adapters
