+++ back/config/services_acceptance.yaml
@@ -4,10 +4,10 @@ services:
         autoconfigure: true
         public: true
 
-    App\Infrastructure\Repository\InMemory\UserRepository: ~
-    App\Domain\Repository\UserRepositoryInterface: '@App\Infrastructure\Repository\InMemory\UserRepository'
+    App\Infrastructure\Persistence\InMemory\Repository\UserRepository: ~
+    App\Domain\Repository\UserRepositoryInterface: '@App\Infrastructure\Persistence\InMemory\Repository\UserRepository'
 
-    App\Infrastructure\QueryFunction\InMemory\GetUserListFromMemory: ~
-    App\Domain\QueryFunction\GetUserList: '@App\Infrastructure\QueryFunction\InMemory\GetUserListFromMemory'
+    App\Infrastructure\Persistence\InMemory\QueryFunction\GetUserListFromMemory: ~
+    App\Domain\QueryFunction\GetUserList: '@App\Infrastructure\Persistence\InMemory\QueryFunction\GetUserListFromMemory'
 
     App\Application\Query\GetUserListHandler: ~
