+++ back/config/services_test.yaml
@@ -4,8 +4,8 @@ services:
         autoconfigure: true
         public: true
 
-    App\Infrastructure\Repository\Doctrine\UserRepository: ~
-    App\Domain\Repository\UserRepositoryInterface: '@App\Infrastructure\Repository\Doctrine\UserRepository'
+    App\Infrastructure\Persistence\Doctrine\Repository\UserRepository: ~
+    App\Domain\Repository\UserRepositoryInterface: '@App\Infrastructure\Persistence\Doctrine\Repository\UserRepository'
 
-    App\Infrastructure\QueryFunction\Doctrine\GetUserListFromDatabase: ~
-    App\Domain\QueryFunction\GetUserList: '@App\Infrastructure\QueryFunction\Doctrine\GetUserListFromDatabase'
+    App\Infrastructure\Persistence\Doctrine\QueryFunction\GetUserListFromDatabase: ~
+    App\Domain\QueryFunction\GetUserList: '@App\Infrastructure\Persistence\Doctrine\QueryFunction\GetUserListFromDatabase'
