+++ back/src/Infrastructure/Persistence/InMemory/Repository/UserRepository.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace App\Infrastructure\Repository\InMemory;
+namespace App\Infrastructure\Persistence\InMemory\Repository;
 
 use App\Domain\Model\Write\User;
 use App\Domain\Repository\UserRepositoryInterface;
